requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Werkzeug==2.3.7
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.20
mysql-connector-python==8.0.33
python-ldap==3.4.3
ldap3==2.9.1
pyad==0.6.0
